+++ homeassistant/helpers/automation.py
@@ -3,7 +3,7 @@
 from collections.abc import Callable, Mapping
 from dataclasses import dataclass
 from enum import Enum
-from typing import Any
+from typing import Any, Final
 
 import voluptuous as vol
 
@@ -13,6 +13,8 @@ from homeassistant.core import HomeAssistant, split_entity_id
 from . import config_validation as cv
 from .entity import get_device_class_or_undefined
 from .typing import ConfigType
+
+CONF_UNIT: Final = "unit"
 
 
 class AnyDeviceClassType(Enum):
@@ -163,3 +165,21 @@ def _validate_number_or_entity(value: dict | float | str) -> float | str:
 number_or_entity = vol.All(
     _validate_number_or_entity, vol.Any(vol.Coerce(float), cv.entity_id)
 )
+
+
+def validate_unit_set_if_range_numerical[_T: dict[str, Any]](
+    lower_limit: str, upper_limit: str
+) -> Callable[[_T], _T]:
+    """Validate that unit is set if upper or lower limit is numerical."""
+
+    def _validate_unit_set_if_range_numerical_impl(options: _T) -> _T:
+        if (
+            any(
+                opt in options and not isinstance(options[opt], str)
+                for opt in (lower_limit, upper_limit)
+            )
+        ) and CONF_UNIT not in options:
+            raise vol.Invalid("Unit must be specified when using numerical thresholds.")
+        return options
+
+    return _validate_unit_set_if_range_numerical_impl
